+++ ChangeLog.md
@@ -16,6 +16,7 @@ Emperor 0.9.4-dev (changes since Emperor 0.9.4 go here)
 * Discrete colors are provided by [colorbrewer](http://colorbrewer2.org).
 * The rainbow colormap has disappeared from emperor, see [this paper](http://ieeexplore.ieee.org/xpl/articleDetails.jsp?arnumber=4118486) if you want to use that colormap.
 * Remove discrete/continuous coloring checkbox and replace for the color-scheme drop down menu.
+* Add option to select QIIME colors from the color-scheme drop down menu.
 
 Emperor 0.9.4 (10 Sept 2014)
 ----------------------------